+++ src/app/component/shared/components/calendar-base/calendar-base.component.html
@@ -1,0 +1,1 @@
+<!-- <p>calendar-base works!</p> -->
