+++ src/app/component/shared/components/calendar-base/calendar-image.ts
@@ -1,0 +1,4 @@
+export const calendarImage = {
+    arrowNext: 'assets/img/profile/calendar/arrow_next.svg',
+    arrowPrevious: 'assets/img/profile/calendar/arrow_previous.svg'
+  };
