+++ src/app/component/shared/components/index.ts
@@ -1,3 +1,4 @@
 export { CancelPopUpComponent } from './cancel-pop-up/cancel-pop-up.component';
 export { EditPhotoPopUpComponent } from './edit-photo-pop-up/edit-photo-pop-up.component';
 export { PhotoUploadComponent } from './photo-upload/photo-upload.component';
+export { CalendarBaseComponent } from './calendar-base/calendar-base.component';
