+++ src/app/component/user/components/habit/add-new-habit/habit-calendar/habit-calendar.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { HabitCalendarComponent } from './habit-calendar.component';
+
+describe('HabitCalendarComponent', () => {
+  let component: HabitCalendarComponent;
+  let fixture: ComponentFixture<HabitCalendarComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ HabitCalendarComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(HabitCalendarComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
