+++ src/app/component/user/components/habit/add-new-habit/habit-calendar/habit-calendar.component.ts
@@ -1,0 +1,27 @@
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { LanguageService } from '@language-service/language.service';
+import { TranslateService } from '@ngx-translate/core';
+import { CalendarBaseComponent } from '@shared/components';
+
+@Component({
+  selector: 'app-habit-calendar',
+  templateUrl: './../../../profile/calendar/calendar.component.html',
+  styleUrls: [
+    './../../../profile/calendar/calendar.component.scss',
+    './habit-calendar.component.scss']
+})
+
+export class HabitCalendarComponent extends CalendarBaseComponent implements OnInit, OnDestroy {
+
+  constructor(public translate: TranslateService,
+    public languageService: LanguageService) {
+    super(translate, languageService);
+  }
+
+  ngOnInit() {
+    this.bindDefaultTranslate();
+    this.subscribeToLangChange();
+    this.buildCalendar();
+  }
+
+}
