+++ src/app/component/user/components/habit/add-new-habit/habit-progress/habit-progress.component.html
@@ -1,1 +1,2 @@
-<p>habit-progress works!</p>
+
+<app-habit-calendar></app-habit-calendar>
